Analysing the spacial possibilites
So my first pass is just to break down the spaces and see what is a realistically possible, it seems the spaces are quite big, and there is enoug room to make pretty sizable rooms, would you say probablly an addition of a small office/room? cuz it seems if you add a 4th room it could be interesting. At this level i'm not satisfied with this but it is just a means of synching some spacial volumes in my mind first. My intention is to come up with atleast 2-3 schemes
Some Requirements by owner:
- Keep the interior first floor configuration
- No moving of concrete walls
- Find a way to put Stairs in
- 3 bedroom 2 bathrooms or one Jack and Jill Bath
- No balconies
- 1 Closet for Mechanical such as furnace
- Design overhang off back 2 give more of a covered Entry Way into exist Kitchen.
- 1 Option they want is having a Large family room addition on the first floor level behind the kitchen.
- I'm afraid this would put them way over budget
- Single gable roof with 8ft walls for second story,
- No occupied attic space
- Keep the roof about 5/12inch overhang and 6inch fascia.
- Single hung windows